simplify evaluation form layout
previously it was possible to view the evaluation form either

1) side-by-side with the submission form
2) with the submission form maximised (by going to a new URL)
3) with the evaluation form maximised (by going to a new URL)

Navigating between each view is done with links disguised as buttons.
This makes the user flow quite confusing as I have observed users
clicking on one of the buttons and being completely confused when
presented with the maximised submission or evaluation form (with
even more cryptic buttons to navigate to the other views).

This commit makes the side-by-side view the default and only possible
state the evaluation form can be in. If a user wants to 'maximise' a
side he can simply drag the slider to resize the pane accordingly. This
makes it simpler for the developers to implement and also simpler for
the users to understand.