added some exploratory template code
Going to try rendering all templates on startup instead of rendering
them on the fly. That requires changing all stateful template functions
(that take in the current request) into template data variables, which
will leave the code in quite a broken state until I complete the
refactor. This is just a checkpoint commit while everything still works
before I attempt to refactor in the next commit.